* 6.0:
  [PhpUnitBridge] Fix test
  [RateLimiter] Add typecase to SlidingWindow::getExpirationTime
  Bump Symfony version to 6.0.8
  Update VERSION for 6.0.7
  Update CHANGELOG for 6.0.7
  Bump Symfony version to 5.4.8
  Update VERSION for 5.4.7
  Update CHANGELOG for 5.4.7
  Bump Symfony version to 4.4.41
  Update VERSION for 4.4.40
  Update CONTRIBUTORS for 4.4.40
  Update CHANGELOG for 4.4.40
  [Messenger] reset connection on worker shutdown
  reset Twig form theme resources between requests

+++ src/Symfony/Component/RateLimiter/Policy/SlidingWindow.php
@@ -60,7 +60,7 @@ final class SlidingWindow implements LimiterStateInterface
      */
     public function getExpirationTime(): int
     {
-        return $this->windowEndAt + $this->intervalInSeconds - microtime(true);
+        return (int) ($this->windowEndAt + $this->intervalInSeconds - microtime(true));
     }
 
     public function isExpired(): bool
